+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Multiline: 
+  - common metrics settings
+  - metrics settings from source data
 - Data-picker: Added the ability to toggle between range and exact date.
 - Select: Added option to clear the value